test: add 262 new tests — unit tests, widget tests, integration tests
Unit tests (209):
- ZATCA: compliance checker (47), offline queue (25), invoice service (13),
  reporting API (10), CSID onboarding (12)
- POS: invoice service (14)
- Sync: conflict resolver (33)
- Core: PIN validation (20), sync queue (35)

Widget tests (53):
- POS: pos_screen (4), payment_screen (4), receipt_screen (5)
- Shared UI: dashboard (5), invoices (9), invoice_detail (4),
  orders (11), order_tracking (5), customers (6)

Also fixes test factory constructors across 6 packages to match
updated ProductsTableData/SalesTableData schemas.

+++ packages/alhai_sync/test/helpers/sync_test_helpers.dart
@@ -219,7 +219,7 @@ StockDeltasTableData createStockDelta({
   String productId = 'product-1',
   String storeId = 'store-1',
   String? orgId,
-  int quantityChange = -3,
+  double quantityChange = -3.0,
   String deviceId = 'device-1',
   String operationType = 'sale',
   String? referenceId,
